Move "Merge Realm&API" methods to their respective Controllers
